ml-training/requirements.txt
tensorflow>=2.12
tf2onnx>=1.14.0
onnx>=1.15.0
onnxruntime>=1.18.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1
shapely>=2.0.0
rasterio>=1.3.8
scikit-image>=0.21.0
PyYAML>=6.0.1
tqdm>=4.66.0
matplotlib>=3.8.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-timeout&gt;=2.2.0
# Sprint 3 Yield RF deps (CPU-friendly)
scikit-learn>=1.2.0
joblib>=1.3.0
pandas>=2.0.0
skl2onnx>=1.15.0
netCDF4>=1.6.4
h5py>=3.9.0
# Optional dependency for decoupled weight decay (AdamW/SGDW) via tensorflow-addons.
# Install to enable native decoupled WD; otherwise the code falls back to L2 kernel regularization.
# tensorflow-addons>=0.20.0  # optional